# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionCreate": {
        "properties": {
          "database": {
            "description": "Postgres: the database name. SQLite and DuckDB: the path of the file",
            "maxLength": 255,
            "minLength": 1,
            "title": "Database",
            "type": "string"
          },
          "dialect": {
            "enum": [
              "postgres",
              "sqlite",
              "duckdb"
            ],
            "title": "Dialect",
            "type": "string"
          },
          "host": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Postgres only",
            "title": "Host"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "description": "Stored encrypted; needs MLPILOT_SECRET_KEY. Never returned",
            "title": "Password"
          },
          "password_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Name of an environment variable that holds the password (nothing secret is stored)",
            "title": "Password Env"
          },
          "port": {
            "anyOf": [
              {
                "maximum": 65535.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Postgres only (default 5432)",
            "title": "Port"
          },
          "ssl_mode": {
            "anyOf": [
              {
                "enum": [
                  "disable",
                  "prefer",
                  "require",
                  "verify-ca",
                  "verify-full"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Postgres only; default prefer",
            "title": "Ssl Mode"
          },
          "ssl_root_cert": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Postgres only",
            "title": "Ssl Root Cert"
          },
          "username": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Postgres only",
            "title": "Username"
          }
        },
        "required": [
          "name",
          "dialect",
          "database"
        ],
        "title": "ConnectionCreate",
        "type": "object"
      },
      "ConnectionRead": {
        "properties": {
          "can_write": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "From the last test: could the role write? None until tested",
            "title": "Can Write"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "database": {
            "title": "Database",
            "type": "string"
          },
          "dialect": {
            "enum": [
              "postgres",
              "sqlite",
              "duckdb"
            ],
            "title": "Dialect",
            "type": "string"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_tested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tested At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Where the password lives: 'env:NAME' or 'stored (encrypted)'; never its value",
            "title": "Secret"
          },
          "ssl_mode": {
            "anyOf": [
              {
                "enum": [
                  "disable",
                  "prefer",
                  "require",
                  "verify-ca",
                  "verify-full"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Mode"
          },
          "ssl_root_cert": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Root Cert"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "id",
          "project_id",
          "name",
          "dialect",
          "host",
          "port",
          "database",
          "username",
          "ssl_mode",
          "ssl_root_cert",
          "last_tested_at",
          "created_at"
        ],
        "title": "ConnectionRead",
        "type": "object"
      },
      "ConnectionTestResult": {
        "properties": {
          "can_write": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "True when the role could modify data",
            "title": "Can Write"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "auth_failed, host_unreachable, timeout, ssl_required, ...",
            "title": "Error Code"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Short and free of secrets",
            "title": "Message"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "privilege_notes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Privilege Notes",
            "type": "array"
          },
          "server_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Version"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ConnectionTestResult",
        "type": "object"
      },
      "ConnectionUpdate": {
        "description": "Every field optional; a password given here replaces the saved secret.",
        "properties": {
          "database": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database"
          },
          "host": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "password": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "password_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password Env"
          },
          "port": {
            "anyOf": [
              {
                "maximum": 65535.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "ssl_mode": {
            "anyOf": [
              {
                "enum": [
                  "disable",
                  "prefer",
                  "require",
                  "verify-ca",
                  "verify-full"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Mode"
          },
          "ssl_root_cert": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Root Cert"
          },
          "username": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "title": "ConnectionUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agentic ML experimentation platform",
    "title": "MLPilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/connections/": {
      "get": {
        "operationId": "list_connections_api_v1_projects__project_id__connections__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectionRead"
                  },
                  "title": "Response List Connections Api V1 Projects  Project Id  Connections  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Connections",
        "tags": [
          "connections"
        ]
      },
      "post": {
        "operationId": "create_connection_api_v1_projects__project_id__connections__post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Connection",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/v1/projects/{project_id}/connections/{connection_id}": {
      "delete": {
        "operationId": "delete_connection_api_v1_projects__project_id__connections__connection_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Connection",
        "tags": [
          "connections"
        ]
      },
      "get": {
        "operationId": "get_connection_api_v1_projects__project_id__connections__connection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Connection",
        "tags": [
          "connections"
        ]
      },
      "patch": {
        "operationId": "update_connection_api_v1_projects__project_id__connections__connection_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Connection",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/v1/projects/{project_id}/connections/{connection_id}/test": {
      "post": {
        "description": "Connect read-only, report the server version and what the role could write.",
        "operationId": "test_connection_api_v1_projects__project_id__connections__connection_id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Connection",
        "tags": [
          "connections"
        ]
      }
    },
...
  }
}